# Conflicts:
#	src/Jobs/Corporation/MembersLimit.php
#	src/Jobs/Killmails/Corporation/Detail.php
#	src/Jobs/Killmails/Corporation/Recent.php
#	src/Jobs/Market/Corporation/Orders.php
#	src/Jobs/Wallet/Corporation/Balances.php
#	src/Jobs/Wallet/Corporation/Journals.php
#	src/Jobs/Wallet/Corporation/Transactions.php
#	src/Models/Corporation/CorporationMembersLimit.php
#	src/Models/Market/CorporationOrder.php
#	src/Models/Wallet/CorporationWalletBalance.php
#	src/Models/Wallet/CorporationWalletJournal.php
#	src/Models/Wallet/CorporationWalletTransaction.php
#	src/database/migrations/2018_01_08_162711_create_character_industry_jobs_table.php
#	src/database/migrations/2018_01_19_213044_create_corporation_wallet_journals_table.php
#	src/database/migrations/2018_01_19_213119_create_corporation_wallet_transactions_table.php

+++ src/database/migrations/2018_01_05_103303_create_mail_headers_table.php
@@ -19,7 +19,7 @@ class CreateMailHeadersTable extends Migration
             $table->bigInteger('character_id');
             $table->bigInteger('mail_id');
             $table->string('subject');
-            $table->integer('from');
+            $table->bigInteger('from');
             $table->dateTime('timestamp');
             $table->json('labels');
             $table->boolean('is_read')->default(false);
